[FIRRTL] Rewrite the FIRRTL to RTL pass completely to be a normal MLIR pass, instead of using the operation legalization framework.
The op legalization framework isn't really set up for cases where you change
types like we are.  The lowering we are doing is super simple, so just do
a linear pass over the ops.  This also makes the code nicer to work with
because we can use a standard visitor instead of registering rewrite patterns.