Update LLVM version to c89e46e76 (#63)
* Update LLVM version to c89e46e76

A small fixup is required to avoid conflicts between ODS-generated build
methods because of default arguments.  Also, since this keeps failing, I've
updated the llvm cache so that it runs in a separate job.  Otherwise the cache is not updated when the CIRCT build fails.

* [LLHD] Add `SigType` conversion

* [LLHD] Add time type and constants conversion

Previously time constants were converted by their users, by creating a constant for each of the three time attributes. This makes time constants LLVM arrays containing the three time values instead, allowing more a more flexible use (e.g. passing time operands as block arguments).
* Add time type conversion.
* Add tme constants conversion.
* Add time operands to the process persistence state.
* Fix affected tests.

* [LLHD] Set LLVM::DialectCastOp as legal during `InstOp` lowering

This works around the PartialConversion failing because of the introduction of (unused) `DialectCastOps`, that will be dealt with later during the full conversion.

Co-authored-by: rodonisi <simon@rodoni.ch>