Fix pointer value persistence
If a value was stored to a pointer in the same block the pointer was defined, the value update would not reflect in the persistence state. This redirects uses of pointers in the same block as their defining op to a pointer in the persistence state directly, as happens for uses outside of the defining block.