[FIRRTL] Fix tests so "main" matches a module
Fix FIRRTL tests to make them valid FIRRTL IR. Previously, the example
FIRRTL IR or FIRRTL MLIR dialect examples could contain a circuit
whose name (what FIRRTL IR calls the "main" or, equivalently, the top
module) did not match the name of a module in the circuit. This commit
changes the tests so that this invariant holds.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>