[FIRRTL] Test circuit verifier
Add a test for the FIRRTL dialect that the user is provided with an
error if they provide FIRRTL MLIR whose name (it's "main" or top
module) does not match the name of any modules in the circuit.

Add a test that the FIRRTL dialect circuit op has a non-empty name.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>