Fix unexpected trace entries and possible instance executions
Multiple drives to the same signals could erroneously trigger trace entries to be printed, and, in the worst case, unexpected instances to wakeup. This should fix the issue by processing one signal at a time, instead of one change at a time, triggering wakeups and trace entries only if the final result for the event represents a change.