[FIRRTL] Add ExtModule verification
Adds verification of the FIRRTL Circuit operand to reject circuits
with invalid FExtModules. This checks the following things:

1. No FExtModule defname conflicts with an FModule symbol name

2. All FExtModules with the same defnames have the same number of
ports

3. All FExtModules with the same defnames have ports with the same
name in the same order.

4. All FExtModules with the same defname have exactly the same type of
port including widths if they take no parameters and exactly the same
type of port excluding widths if they take parameters.

Adds tests of the above behavior.

This aligns FIRRTL Dialect behavior to accept and reject the same
circuits that the Scala FIRRTL Compiler accepts and rejects in its
test suits that involve ExtModules.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>