[LLHD] Add most missing LLVM dialect lowerings (#151)
* [LLHD] Convert and persist block arguments in process lowering

* [LLHD] Add full `WaitOp` lowering and simulation

* Complete the `WaitOp` lowering by also lowering sensed signals.
* Fix conversion failing when the optional time operand is missing.
* Fix resume indexes getting mixed up.
* Add expected wakeup time to process simulator state, to invalidate scheduled waits that have already been triggered by a sensed signal.

* Persist subsignals

* Fix process entry block; wait destination block arguments

* Fix pointer value persistence

If a value was stored to a pointer in the same block the pointer was defined, the value update would not reflect in the persistence state. This redirects uses of pointers in the same block as their defining op to a pointer in the persistence state directly, as happens for uses outside of the defining block.

* Fix proc senses malloc size

* MLIR update fixes

* [LLHD] Lower inss to LLVM

* Add lowering pattern for inss, only for integers for now.

* Gather integer from LLVMTypes too

This allows to gather operands widths more consistently, in cases where
an already converted operand is received.

* Fix insertions for non-zero start

* Remove comment artifact

* Fix mask width

* Add inss conversion test case

* [LLHD] Lower DrvOp enable to LLVM

This completes the `DrvOp` lowering by adding the lowering of the enable condition. This was previously ignored.

* [LLHD] Add `RegOp` lowering

* Add `RegOp` lowering to LLVM.
* Keep track of the trigger values for the `fall`, `rise` and `both` modes, as those require comparison with previous cycles.
* Change the wakeup queue to a set. If multiple signals owned by the same instance change, the instance got queued more than once per cycle, resulting in unexpected behavior. A set should ensure this doesn't happen.

* Adjust state destructor

* Use smallset for the wakeup queue

* Fix same-name instances

Instances with the same name but instantiated by different entities
would previously conflict with each other due to them being mappend only
by name. Using the instantiating unit name in the mapping keys should
solve the conflicts, as the instance names are unique within an entity.

* Fix malloc size for process name string

This was not considering the null terminator for the string, resulting in writes outside the malloc'd region

* Get correct reg trigger index

* Add reg conversion test

* [LLHD] Change simulator time granularity to ps

* Previously only the time unit was ignored, and time delays were assumed to be in `ns`. This makes the base unit for the simulator be `ps` and considers the time unit when lowering.
* Promote the simulator's state time values to uint64_t to accomodate the new base time unit.

* Promote lowered time value to i64

Converting `s` to `ps` would overflow otherwise.

* Fix tests

* [LLHD] Lower array type and initialization.

* [LLHD] Lower array signals

* Add array conversion tests

* Get drvOp type from value; cover case of already converted array

* Fix integer dump widths

* Covnert extraction operations on arrays

* Fix alloca type

* Properly handle bitwidth adjustments

Previously integer bitwidth adjustments were mostly handled with arbitrary decisions. This gives more flexibility, by dynamically adjusting the value to the desired type.

* Merge integer slice extraction duplicate code into a function

* Move constant creation to more appropriate location

* Add array extract/insert conversion tests; move extract/insert tests to
separate files

* Use make_pair

* Redirect signal and array shr to dyn_extract

Temporary workaround until we get rid of those shift instructions

* Fix ShrOp redirect case when an operand type is already converted

* [LLHDToLLVM] Lower tuple type, init, extracts

* Fix gep indexes for structs

This ensures indexes for struct gep are legal 32 bit constants.

* Add tuple conversion tests

* Fix drv width computation

* Check for converted type in DrvOp conversion too

In some cases, the DrvOp conversion receives an already converted type, notably when generated from a RegOp. This should ensure the conversion still goes through without problems.

* Add array boundary check for dyn_extract_slice

Dynamic extracts are expected to still work when accessing out of bounds, returning the base value for the type. This adds this check for dynamic extracts on arrays, returning a 0-valued element for each of the elements outside of the array's boundary.

* Also check for LLVM structured types in extf lowering

* Add NegOp conversion

* Lower equality operations

* Fix simulator failing on instances with the same signal passed multiple times

So far the simulator implementation assumed signals to be passed at most once as argument to an instance. This is expected to be legal though, and especially happens on signals appearing as both input and output.

* Add missing lines at end of files

* Use pow for time unit transformation

* Adjust comments

* Move pointer redirect to existing loop

* Use existing variable

* Move sparse helpers to helper section

* Fix missing return; add boundary check simulation test

* Refactor variable name

* Add entitiy init comments

* Directly replace op

* Use mul for negation

* Add time constant comments

* Fix NegOp test

* Clarify comment

* Make runtime functions camelcase

This should make clang-tidy happy. Updated all the runtime functions for consistency.

* Directly pass value to getI32ArrayAttr

* Merge getProcPersistenceTy ptr and sig cases

* Refactor getSigType -> getLLVMSigType

* Split convert_process test in more manageable chunks

* Make global mutable state local to runOnOperation

* Promote entity and process local state to unique_ptr

* Promote state signal values to unique_ptrs

* Missing newline at end of file