[ESI] [Cosim DPI] RTL
This is the fourth in a series to merge #213 -- RTL. Adds the
SystemVerilog code which will interface with the DPI library.

Is this the right place for SystemVerilog distributables? I put them
here since the 'include' directory is shipped with the install. Should
we have a top level "support" directory which goes into /usr/share?