[LowerToRTLModule] Fix a bug lowering modules with outputs before inputs.
We were walking the port list in the order of the RTL module, not the
order of the FIRRTL module as we needed to.  The later always produces
inputs before outputs and we needed the order of the source module.

This manifested itself as a variety of "error: source and result width
must match" errors from the verifier.