[RTL] Add support for hard coded 'verilogName' attribute on extmodules
The RTL dialect doesn't support proper parameterized types yet, so we
need to have multiple declarations for parameterized external modules
with the various type signatures that are used.  These need different
symbols, but need to end up being the same verilog module name.

Borrow a trick from FIRRTL where we support a "verilogName" attribute
that allows us to hard code this.  This should go away when we support
proper parameterized types, but will allow us to make progress in the
meantime.