[LowerToRTL] Improve output lowering logic to avoid temporary wires.
This is a very narrow implementation of a general optimization in
LowerToRTLModule: when lowering outputs, look to see if there is
one connect that sets it in FIRRTL.  If so, use that value instead
of introducing a temporary wire.

FIRRTL has a bunch of weird cases to support, so this includes a
testcase that explores these.  This patch doesn't implement the
set of exciting things though.