[llhd-sim] Move trace generation to a dedicated class (#229)
* Add trace format interface

This adds a new class that makes it easier to add new dump formats (includes a couple of variations of the trace format used so far) to the simulator. This fully moves trace generation to the new itnerface, which also fixes:
* The signals within a time-step are properly sorted by their paths.
* Signals of arrays and tuples generate one trace entry for each of their element, instead of being printed as ints.

* Fix make_pair

* Parse mlir files through a SourceMgr

This changes how the mlir source files are parsed, making it more similar to what is done in MLIROptMain. This also enables piping a module into llhd-sim directly, which was previously not possible.

* Add option to stop simulation at a real-time step.

* Use sort vector + erase-unique  for wakeupQueue

* Add named-only trace format

This introduces a new merged trace format that filters out default-named signals with name 'sig[0-9]*'. This is useful to compare the trace against converted vcd traces of the top-level signals.

* Remove unused variable

* Adapt signal name filtering for named-only trace

* Update completition message

* Rebase fixes

* clang-format

* Debug artifact

* Merge redundant change push code

* Fix typo

* Fix root instance map key

* Remove unnecessary map entry

* Refactor redundant code

* Simplify if statement; comments

* Define traced signals only once on initialization

* Don't trace structured signal elements that do not change, for all formats

* Refactor sort in a private function

* Add boilerplate headers

* Ensure nothing is printed when there is no change

* Add formats test

* Minor adjustments/comments

* Simplify `lastValue` map key

* Adjust formats descriptions

* clang-tidy

* Skip clang-tidy header-guard check

* Comment typo

Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>

Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>