[FIRRTL] Allow firrtl.instance in the body of firrtl.when
firrtl.instance must be enclosed by a firrtl.module, but it may not be
the direct parent op. When verifying, using parentOfType will search all
parents of the firrtl.instance op during verification.  This allows a
firrtl.instance op to appear in the body of ops other than a module,
such as firrtl.when.

firrtl.when is still not handled during lowering to RTL.  Any
firrtl.instance occuring in the body will not be properly lower to an
rtl.instance.  This will cause verification to fail.