[FIRRTL] Fix Or canonicalization (#312)
The `or` operation canonicalizes `or(x, x) => x`.  The problem is that
when the output type doesn't match the input type, this transformation
is invalid and causes a type mistmatch.  This change prevents invalid
folding, and adds a similar test case for `and`, `xor`, and `or`.  In
the future this should canonicalize to a `cast` operation.