[LowerToRTL] Fix a dangling pointer bug in LowerToRTL.
CI started failing in a way I can't reproduce.  Upon inspection, it
looks like the problem is that CI getting "lucky" at finding a bug
that would be impossible to track down otherwise.

The issue is that the valueMapping dictionary is getting populated
with keys in one module, but those keys never get removed as the
pass moves on to other modules.  If IR objects happen to get allocated
at the same address, they'll show up as ghost entires in the map,
causing all sorts of bad things to happen.

Fix this by clearing the map at the end of the pass.  There is no
reasonable way to write a testcase for this unfortunately.