[FIRRTL] Tighten up firrtl.module port requirements to require FIRRTL types
For awhile there we were allowing standard types and other things as firrtl
module ports.  While the RTL dialect is still permissive in what it accepts,
there is no reason to support this in the firrtl dialect.

Most of this is updating testcases that were using a mix of things.