[Handshake] Update created InstanceOp types to match modules.
The FIRRTL spec declares that the bundle type used to represent an
instance has fields named the same as the ports of the module. It
turns out this property is being relied on by the LowerToRTL pass in
the FIRRTL dialect.

This code in the HandshakeToFIRRTL pass has been generating names for
all fields in the instance. That actually works fine, except when the
specially named "clock" and "reset" ports are used. The update ensures
that the generated instance's type has the same names as the module
ports.