[RTL] Add more constant folding
It is important to perform constant folding in an Op's folder, not in
canonicalization patterns.  This is because the Op's folder is invoked
in multiple places, such as in SCCP for Op interpretation.  This change
adds constant folding functionality to all the variadic operations'
folders. The end effect is that some of the functionality is duplicated
between the constant folder and canonicalizer.

It is not possible to combine the constant operands into a single
operand during constant folding, and so this must be handled during
canonicalization. This is due to the limitiation that the folder cannot
create new operations, even if they are constant.
```mlir
// not possible during constant folding:
and(x, 5, 3) -> and(x, 1)
```

It is possible to change the number of arguments during constant
folding, so it would be possible to perform operand de-dup for the
set-like operations (`and` and `or`). This is not covered.