[FIRRTL] Move transformation into a transformation library
We will eventually adding some more transformation passes to FIRRTL, so
I thought that it would be a good idea to move them into their own
library. A couple upgrades were made on the way.

ODS pass registration was moved into a headfile, to be more inline with
MLIR.  This means we no long have to manually create a pass registration
function.  There seems to be two common choices for naming the pass
registration, the second one was taken:

```c++
void registerFIRRTLPasses();
void firrtl::registerPasses();
```

Without having to worry about pass registration, I was able to delete
DialectRegistration.cpp.

One thing which was not totally clear was where the `Passes.h` header
file should live: in `FIRRTL` or `FIRRTL/Transfoms`.  There seems to be
a pretty even split in MLIR.  I put Passes.h `FIRRTL/Transforms` since
users are required to link against the Transforms library to use the
functions.

The pass `lower-firrtl-types` was renamed to `firrtl-lower-types` under
the idea that all dialect specific passes should be namespaced by the
dialect. We could also have also named the pass
`firrtl-lower-firrtl-types` if that makes more sense.