[FIRRTL] Replace the firrtl.invalid stmt with a firrtl.invalidvalue expression.
This makes the FIRRTL dialect easier to work with (e.g. the "merge" inference
algorithm in LowerToRTLModule works with invalid now), allows more wire
promotion, opportunities and doesn't affect expressive potential.

This is a deviance from the design explained in the FIRRTL spec so I added a
new section to the rationale to describe it.