[FIRRTL] Add division folds (#405)
This adds folding to the FIRRTL dialect division operand. Two folds
are implemented:

  - div(a, a) -> 1
  - div(a, 1) -> a

The former exploits the fact that the FIRRTL specification allows
division by zero to be undefined. This is then used to treat division
by zero as 1 (which is the same as any division by self) and produce a
constant.

The latter is implemented for UInts only. SInt division is width
expanding which means that a pad operand should be used (which puts
this canonicalization outside the realm of a fold which cannot add
operands).

Tests and an updated description are included.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>