[HandshakeToFIRRTL] Add MemoryOp lowering. (#417)
* [HandshakeToFIRRTL] Add MemoryOp lowering.

This adds a basic lowering from MemoryOp. The appropriate
data signals are connected and the control signals are built
similarly to the ForkOp and JoinOp. The memory uses a read
latency of 0, a write latency of 1, and returns the old value for read
under write. This is part of https://github.com/llvm/circt/issues/337.