[FIRRTL] Return multiple results from memories
Change FIRRTL Dialect's MemOp to return one result per port as opposed
to a single result bundle containing each port as a subelement.

Fix bug where a memory with a read/write port would erroneously return
null from getDataTypeOrNull. This will now use either "data" or
"rdata" to look up the field that has the data type of the memory
based on the kind of the first memory port.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>