[RTLToLLHD] Rework RTLModuleOp conversion to use upstream pattern. (#499)
The upstream dialect conversion pattern for rewriting a function type
has been updated to support all FunctionLike ops, and we can use it
here. This removes all of the signature and region conversion from the
pattern, and replaces it with a simple check that the conversion
succeeded. Because the upstream conversion doesn't turn outputs into
inputs, we also need to add all of the RTLModuleOp outputs as inputs
to the EntityOp.