[SV] Add pass to merge AlwaysFF operations
This adds a pass to merge `sv.alwaysff` operations with identical
conditions.  Due to the design of this operation it is trivial to check
if we meet the conditions for merging this operation.  This pass is
structured to not depend on the RTL dialect, and does not assume that
the `sv.alwaysff` operations are in an RTL module.  They do, however,
need to be in a graph region to work.

Although `sv.alwaysff` regions are not graph regions, they are limited
to having a single block.  This pass takes advantage of this to quickly
splice blocks together when merging operations.