[ExportVerilog] Issue#439: Collect names recursively under scopes.
This makes sure we emit wire declarations for expressions that have
multiple uses correctly, avoiding an exponential blowup of verilog
in some cases, and emitting expressions that have to go through an
intermediate wire declaration.

This is currently emitting wire declarations within always blocks,
which I'm not sure is valid.  If not, I'll do a second follow up.