[ExportVerilog] Emit multi-use SSA values as "logic" declarations instead of wire.
It turns out these are completely different forms in the SV spec (wire is a net,
logic is a var decl).  logic corresponds to what we actually need (single assignment
semantics) and can be used in nested blocks like an always, so we use it instead of
wire generally.

Explicit rtl.wire declarations and rtl.merge are still emitted as a wire of course.

This is progress towards issue #439.