[HandshakeToFIRRTL] Fix decoder one-hot vector width.
The previous calculation was done using the width of the entire 64-bit
index to dynamically shift a bit into a one-hot vector. This adds a
tail operation to only select the bits needed from the index. This
updates the decoder implementation to just compute the correct result
width of the dshl operation, and drops the unneeded pad.