[Comb][RTL] Add rtl.instance inlining support
This adds the DialectInlinerInterface to both the Comb and RTL dialects.

In Comb, this interface is used to specify that all Comb operations are
legal to inline. This means that a Comb operation appearing in the body
of a module will not prevent inlining of the module. For the RTL
dialect, all operations are similarly considered safe to inline.  The SV
dialect has not been considered yet, and all operations are by default
not safe to inline.

The Module and Instance operations in the RTL dialect were marked as
Callee-like and Caller-like using operation interfaces.  All InstanceOps
which have the non-dialect singleton attribute `inline` are considered
safe to inline by the inliner.  This use of this attribute in this way
is a quick hack, and in the future we probably want to use something
more robust.