[SV] Change ifdef ops to have a canonical way to represent ifndef.
Before we'd use "!FOO" as an identifier to get an inverted conditional.
Such a representation gives us two different ways to represent the same
thing and unnecessarily makes transformations on ifdefs (like merging
of them) more complicated.

Instead just put inverted logic into the else block, and have the verilog
exporter print a ifdef with no "then" as ifndef.

This patch also merges the code I duplicated in the verilog emitter for
the two ifdef cases.