[FIRRTL] Lower non-trivial memory latencies to RTL (#585)
Add lowering of memories with non-zero read latency and non-unary
write latency using the strategy of the Scala FIRRTL
Compiler (SFC). If memories have read or write ports with this
property, then add delay pipes that delay the read or write for the
expected number of cycles. This deviates slightly from the SFC
behavior and creates aggregates for the pipes as opposed to one
element for each stage of the pipe.

Remove incomplete support for aggregate memory lowering. Before this
commit, memories would be split, but there was no logic to actually
handle these memories. Change this so that aggregate memories are an
error in lowering and tell the user how to run FIRRTL type lowering.

The existing register initialization logic is extended to support
arrays (as this is required for the pipeline registers).

Update tests and add a new test to check the pipeline register
behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>