[FIRRTL] Candidate fix for #640, NFC
Function argument order is undefined and LowerToRTL creates some
read_inout ops as arguments when creating a comb::AndOp (as the gate
for a MemOp write pipe). Change the lower-to-rtl test to use CHECK-DAG
to allow for these to be legally reordered if a user uses a compiler
which decides to go a different direction with argument order
evaluation.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>