[FIRRTL] Fix unnecessarily repeated calls to setType in LowerTypes.
The previous implementation would call setType to update the module's
type attribute after each new argument was added. This constant
updating of the attribute was needed because getNumFuncArguments used
the attribute, and getNumFuncArguments was called by core MLIR logic
throughout this pass.

By updating getNumFuncArguments to just return the current number of
entry block arguments, we are able to update the type attribute once
at the end of the pass.