[CI] Do not schedule nightly builds on forks
When people fork a repository all of the github-actions are
automatically enabled.  Forked repositorys can manually disable actions
on their fork, but they are by default enabled, and disabling prevents
manually triggering the builds.  If someone needs to run the nightly
builds on their fork this will probably have to be removed.