[CI] Stop building LLVM in a separate workflow job
LLVM is currently built and cached at the beginning of every CIRCT PR.
This is a totally unnecessary step, as we have to pull and build LLVM in
the CIRCT build job as well. Even on a cache hit, this job is still
copying all of the built LLVM files to the build machine.

This separate job was originally added in ad07533 at a time when CIRCT
was not consistently in a buildable state.  When the CIRCT build job
fails, it will not cache the LLVM build directory. To get around this, a
separate job was created to try to populate LLVM in the cache.  Now that
the main branch of CIRCT is stable and passing builds, this job is not
necessary.