[LowerToRTL] Ensure we maintain the firrtl sequential semantics when lowering procedural statements (#676)
The body of always blocks are not ordered with respect to other always blocks. FIRRTL semantics say printf (and friends) have sequential semantics in a module, thus they need to be lowered to a common always block. We can't do this via merging later as that's just best-effort and could result in different orders.