[FIRRTL] Split connect error tests from regular tests
Regular tests cannot be in the same file as error tests, due to the
absence of FileCheck from the diagnostic verification tests.  Before
this change, the positive checks were accidentally not being run.

The error tests were moved into connect-errors.mlir.  The regular tests
are no longer tested using split-input-file, and all modules had to be
renamed to avoid a name collision.