[CI] Add GCC build, combine release and debug builds into release+assert
This change adds a GCC build to the buildAndTest job which is run on
every PR. This step takes about 1m 30s to complete.

This also combines the two clang builds, in release and debug mode, into
a single build, release+asserts. Normally, a cmake release build adds
-DNDEBUG to all command line options. The LLVM cmake system has a flag
to override this, LLVM_ENABLE_ASSERTIONS. Building in release mode with
assertions should give us the best of both worlds for error checking.

The GCC builds will be running in the same release+assert build
configuration.