[RTLCleanup] Fix an N^2 compile time issue in RTLCleanup, NFC.
This speeds up "firtool aliaksei-test1.fir -lower-to-rtl -disable-output"
from 126 to 35s on my machine with a debug build of firtool.

The improvement here is to revisit the subregions of merged nodes (e.g.
ifdefs) once at the end of the sweep over the model, instead of revisiting
the regions after every merge.