[Python] Set up initial Python bindings for CIRCT.
This is the first step towards #710. This should be the minimal amount
of boilerplate to set up a Python module such that we can write
`import circt`. Note that this module does not actually bind to or
expose any API at the moment; this is just the boilerplate, and that
will follow.

The setup here is based on both the upstream MLIR Python bindings, as
well as what NPComp does. When the two diverge (e.g. directory
structure, copying files vs. symlinking), this takes the same approach
as upstream MLIR.