[FIRRTL] Parse local Circuit, Module Annotations
Adds the ability to parse and store CircuitTarget or ModuleTarget
local Annotations as MLIR Attributes. Annotations are stringly-typed
in an array of Dictionary Attributes. The "target" field is removed
as part of this process as this is now implicit based on the operation
to which the Attribute is applied.

Any other types of Annotations (e.g., ComponentTarget or non-local)
produce a runtime error and are promoted to Attributes on the circuit
with the "target" field preserved.

The utility firtool is updated to take an --annotation-file argument
which can be used to import annotations from an annotation file.
Additionally, annotations can be specified using a new "inline" syntax
on a FIRRTL circuit using `%[annotations]`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>