[LowerToRTL] Use a macro in read port lowering to simplify generated Verilog.
Before we'd generate something like:

  `ifdef RANDOMIZE_GARBAGE_ASSIGN	// z.fir:8:5
    assign u0_r0_data = _T_0 < 6'h2A ? _T : `RANDOM;	// z.fir:8:5
  `else
    assign u0_r0_data = _T;	// z.fir:8:5
  `endif

now we generate this at the start of the file:

// RANDOMIZE_GARBAGE_ASSIGN enable range checks for mem assignments.
`ifdef RANDOMIZE_GARBAGE_ASSIGN
  `define RANDOMIZE_GARBAGE_ASSIGN_CHECK(INDEX, VALUE, SIZE) \
    ((INDEX) < (SIZE) ? (VALUE) : `RANDOM)
`else
  `define RANDOMIZE_GARBAGE_ASSIGN_CHECK(INDEX, VALUE, SIZE) (VALUE)
`endif

and then emit the read like this:

  wire [8:0] _T = RANDOMIZE_GARBAGE_ASSIGN_BOUND_CHECK(_T_0, u0[_T_0], 6'h2A);

In addition to tidying up the generated expression, this makes various
cleanup passes work better because the result is only assigned in one
place.  This allows us to eliminate some more extraneous wires.