[ExportVerilog] Add option to control alwaysff printing
This change adds the ability to control the verilog emitter using a set
of attributes on the top level ModuleOp. This is controlled by an
attribute key on the top level module, `circt.loweringOptions`, with a
StringAttr parameter.

This adds a command line option which will amened the IR after parsing
with the options.

The options used in the attribute and the commandline match.