[FIRRTL LowerTypes] Lower bundle wire connects in lower-types (#688)
Since wires have a duplex type, they can appear on the RHS or LHS of a
connect statement, regardless of whether or not they are a passive type.
When lowering a bulk connect statement involving a wire (i.e.
connecting a bundle type), we need to make sure that the non-duplex type
is used to set the direction of the flow. When doing a bulk connect of
two wires, it is considered an ambiguous flow and an error. To connect
two wires, each field must be connected individually, which is
considered a mono-connection with explicit flow from the RHS to the LHS.

This change adds a function to determine if a value comes from an
expression with duplex flow. This property is transitive through
subfield accesses.