[FIRRTL] Parse CircuitTarget and ModuleTarget Annotations (#616)
* [FIRRTL] Parse local Circuit, Module Annotations

Adds the ability to parse and store CircuitTarget or ModuleTarget local
Annotations as MLIR Attributes.  Annotations are stringly-typed and stored
in an ArrayAttr.  The "target" field is removed as part of this process as
this is now implicit based on the operation to which the Attribute is
applied.

Other types of Annotations (e.g., ComponentTarget or non-local) produce a
runtime error.

The utility firtool is updated to take an --annotation-file argument which
can be used to import annotations from an annotation file.  Additionally,
annotations can be specified using a new "inline" syntax on a FIRRTL
circuit using `%[annotations]`.  Any --annotation-file annotations are
appeneded to inline annotations if they exist.

Add tests of CircuitTarget, ModuleTarget, CircuitName, ModuleName, and
erroneous/unsupported annotations or JSON.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>