[Python] Add Python builder for RTLModuleOp.
This is done using MLIR's Python extension approach. This isn't
actually documented at the moment, but the gist is you can add to the
sugar class for a given operation by defining things in a well-known
module.

This particular extension is basically a copy of the builting FuncOp
builder from upstream MLIR, with some RTLModuleOp flavored tweaks. It
is by no means the final state, but this is a decent starting point to
be able to build RTLModuleOps in a somewhat decent way from Python.

See the upstream extension for cross-reference:
https://github.com/llvm/llvm-project/blob/32f6a15dfd5221d5fb628f74ecfe72302af460f4/mlir/lib/Bindings/Python/mlir/dialects/_builtin_ops_ext.py#L25