[ExportVerilog] Use stmt, not expr for temp decl
Pass the current statement around when exporting its expressions.  Use
this statement, not the expression, when generating a declaration
keyword ("reg", "wire", or "automatic logic") for an inlined
expression.  Also, store a pointer to the statement in the ExprEmitter
class.  Use this during retroactive, long line expression emission to
determine if this needs an out-of-line expression declaration.  This
avoids a bug where an expression is inlined from a non-procedural
region to a procedural region and the declaration should change.
Previously, only the expression's defining op would determine this.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>