Support ReferenceTarget Annotations for Operations (#823)
* [FIRRTL] Refactor parser, mutable target tracking

Modify the FIRRTL parser's getAnnotations method to use Twine instead
of StringRef.  This avoids the high cost of string concats.  Add an
early exit if no annotations exist.

Track the circuit and module target in mutable StringRefs.  These are
updated once they are known, i.e., the circuit target is known during
circuit parsing and the module target is known during module or
extmodule parsing.

* [FIRRTL] Custom InstanceOp attr-dict parse/print

Add a custom attr-dict parser/printer specific to InstanceOp.  This
does nothing different for parsing, but for printing always hides
"moduleName" and hides "annotations" if it doesn't exist or is an
empty array.

* [FIRRTL] Accept ReferenceTargets in fromJSON

Relax checking in fromJSON to accept local ReferenceTargets that
contain no subfield or subindices.

* [FIRRTL] Add Annotations to InstanceOp

Add an annotations parameter to InstanceOp that is stored in the
attr-dict.  This has a default value, but not a custom builder with a
default.  Update existing sites that create InstanceOp to pass in a
dummy, empty annotation array.

Use the new custom attr-dict parser/printer for InstanceOp to hide
annotations if they either do not exist or are empty.

* [FIRRTL] FIRScopedParser util to get ModuleTarget

Add a utility to FIRScopedParser that returns a StringRef pointing at
the current ModuleTarget (this lives in private state).  This is
intended to be used to construct more complex targets, e.g.,
ReferenceTargets.

* [FIRRTL] Parse InstanceOp Annotations

Add support to read inline or file-based Annotations that are
associated with local ReferenceTargets that point at an instance.

* [FIRRTL] Add tests of InstanceOp Annotations

* [FIRRTL] rm newly supported ReferenceTarget tests

* [FIRRTL] CMemOp ReferenceTarget Annotations

Add support to attach ReferenceTarget or ComponentName Annotations
pointing at CMemOps to the CMemOps during parsing.

* [FIRRTL] Test CMemOp ReferenceTarget Annotation

* [FIRRTL] SMemOp ReferenceTarget Annotations

Add support to read in ReferenceTarget or ComponentName annotations
targeting SMemOp.

* [FIRRTL] Test SMemOp Annotations

* [FIRRTL] MemOp ReferenceTarget Annotations

Attach ReferenceTarget/ComponentName annotations to FIRRTL MemOps.
During LowerTypes, any annotations are just copied to lowered
memories.

* [FIRRTL] Add MemOp ReferenceTarget Annotation Test

* [FIRRTL] Node ReferenceTarget Annotations

Add support for attaching ReferenceTarget/ComponentName Annotations to
nodes.

Note: since temporary nodes (e.g., _T_0) are dropped by the parser,
any Annotation targeting a temporary is dropped.  Add a "TODO" comment
about this in the associated code in the parser.

* [FIRRTL] Test Node Annotations

* [FIRRTL] Add Wire ReferenceTarget Annotations

Add support for parsing and storing ReferenceTarget/ComponentName
Annotations pointing at wires.

* [FIRRTL] Add tests of Wire Annotations

* [FIRRTL] Reg/RegReset Annotation Support

Add support for ReferenceTarget/ComponentName Reg and RegReset
annotations.

* [FIRRTL] Test Reg/RegReset Annotations

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>