Use ODS generated dialect defintions
In CIRCT there is duplicated declaration of each dialect in ODS and in
C++.  We have been generating the class from ODS, but never including it
anywhere.  This change removes the handwritten dialect classes and uses
the generated version.  In all cases the resulting code is now
equivalent.

Some dialects  with constant materializers had to marked as so in ODS
to have the proper function declaration generated.

In order for the dialects to have the functions to parse types or
attributes, they had to have at least one DialectType or DialectAttr.
This change moved some types and attributes to correctly belong to their
dialect in ODS.