[FIRRTL] Make names non-optional and use names as SSA values
The first part of this change makes names on FIRRTL operations
non-optional StringAttr attributes. Empty names  are still allowed, but
are now represented by an empty string "", instead of a non-existent
attribute.  This change affects the builders generated for each op, and
they now have a builder capable of taking a StringRef for the name.

The second part of this change modifies the SSA printer for registers to
use the same ImplicitSSAName as WireOps.  Since registers now require an
explicit name, making this change now helped keep down the number of
test line changes.

There is a small improvement to the name printer to not print empty
name strings.