[RTL][Python] Add InstanceBuilder for constructing instances.
This adds an InstanceBuilder class to help construct InstanceOps from
Python. The builder accepts a name for the instance and optional
values to assign to the input ports.

The InstanceBuilder supports attribute getters and setters for the
result and input ports, respectively. When all of an InstanceBuilder
input ports are set, either on construction or through a setter, the
instance is created.

A create method is added to RTLModuleOp, which constructs and returns
and InstanceBuilder.