[FIRRTL] Properly apply annotations to entities with temp names
There is a problem where entities with dropped temporary names (such as
"_T_0") can't have annotations properly attached. This change looks up
the annotations of an entity before removing the name, which allows
annotations to apply as intended.

In addition to this, entities with the DontTouch attribute will not have
their name dropped by the parser.