[FIRRTL] Fix `is invalid` on duplex types and bundles
When attached to bundle values, `is invalid` recurses through the bundle
and connects invalid to any flipped field.  When the bundle value is a
register or wire, it should connect invalid to every field, regardless
of the flippedness of that field.  Invalid was not properly recursing or
handling wire types, and this change fixes that.

Due to the way `is invalid` treats wires differently than other values,
we cannot simply bulk connect a bundle typed `invalid`, and let expand
connects handle this later.

Fixes #942