[FIRRTL] Nodes shouldn't be no side effect.
Nodes, especially those marked with DontTouch, can't be removed by dead
code elimination.  Some tools will expect them in the final output.
There is already a canonicalization to remove a node if it has an empty
annotation list, and it should be able to handle this case as well.

This change removes the NoSideEffect trait from the node operation.