[SV] Do not remove dead wires
We have a canonicalization which removes SV wires without any readers.
This is not a legal optimization as wires are considered public and
other tooling may be expecting these wires in the final output.  This
change removes the canonicalization responsible for removing the wire.
This is a useful idea, and we will be revisiting this optimization when
there is a concept of public vs private wires in the RTL dialect.