[SimpleCanonicalize] Implement canonicalize logic from mainline canonicalize pass.
This is the same except it is not calling simplifyRegions, which is irrelevant to
what we're doing in general.  When used on a largish real world example, this cuts
down firtool time of the two canonicalization passes significantly:

  firrtl.module pass: 3.49s -> 2.63
  rtl.module pass: 0.47s -> 0.19s

There are other improvements that I'll land that were reverted from upstream
as well, but this is already a significant win.